Forecast/requirements.txt
xarray>=2024.1.0
cfgrib>=0.9.12.0
pandas>=2.2.0
numpy>=1.26
requests>=2.32.0
python-dotenv>=1.0.1
